requirements.txt
flask
flask-cors
openai
python-dotenv
pandas
gunicorn
